add classes.h to store all the classes and include guard for header files
